# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vortex Hosting — Minecraft server management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ServerSoftware =
  (typeof ServerSoftware)[keyof typeof ServerSoftware];

export const ServerSoftware = {
  paper: "paper",
  leaf: "leaf",
  fabric: "fabric",
} as const;

export type ServerPlan = (typeof ServerPlan)[keyof typeof ServerPlan];

export const ServerPlan = {
  free: "free",
  starter: "starter",
  pro: "pro",
  enterprise: "enterprise",
} as const;

export type ServerStatus = (typeof ServerStatus)[keyof typeof ServerStatus];

export const ServerStatus = {
  stopped: "stopped",
  starting: "starting",
  running: "running",
  stopping: "stopping",
} as const;

export interface Server {
  id: number;
  name: string;
  software: ServerSoftware;
  version: string;
  plan: ServerPlan;
  port: number;
  status: ServerStatus;
  memoryMb: number;
  maxPlayers: number;
  playerCount?: number;
  /** @nullable */
  uptimeSeconds?: number | null;
  createdAt: string;
}

export type ServerInputSoftware =
  (typeof ServerInputSoftware)[keyof typeof ServerInputSoftware];

export const ServerInputSoftware = {
  paper: "paper",
  leaf: "leaf",
  fabric: "fabric",
} as const;

export type ServerInputVersion =
  (typeof ServerInputVersion)[keyof typeof ServerInputVersion];

export const ServerInputVersion = {
  "1214": "1.21.4",
  "1204": "1.20.4",
  "1201": "1.20.1",
  "1194": "1.19.4",
} as const;

export type ServerInputPlan =
  (typeof ServerInputPlan)[keyof typeof ServerInputPlan];

export const ServerInputPlan = {
  free: "free",
  starter: "starter",
  pro: "pro",
  enterprise: "enterprise",
} as const;

export interface ServerInput {
  /**
   * @minLength 1
   * @maxLength 64
   */
  name: string;
  software?: ServerInputSoftware;
  version: ServerInputVersion;
  plan: ServerInputPlan;
  maxPlayers?: number;
}

export interface CommandInput {
  command: string;
}

export interface LogLine {
  id: number;
  line: string;
  timestamp: string;
}

export interface MetricPoint {
  id: number;
  playerCount: number;
  uptimeSeconds: number;
  recordedAt: string;
}

export interface ApiError {
  error: string;
}